+++ src/cue/sectioncuehandler.h
@@ -37,7 +37,7 @@ namespace PixelMaestro {
 			uint8_t* set_dimensions(uint8_t section_num, uint8_t layer_num, uint16_t x, uint16_t y);
 			uint8_t* set_layer(uint8_t section_num, uint8_t layer_num, Colors::MixMode mix_mode, uint8_t alpha);
 			uint8_t* set_offset(uint8_t section_num, uint8_t layer_num, int16_t x, int16_t y);
-			uint8_t* set_scroll(uint8_t section_num, uint8_t layer_num, int16_t x, int16_t y);
+			uint8_t* set_scroll(uint8_t section_num, uint8_t layer_num, uint16_t x, uint16_t y, bool reverse_x = false, bool reverse_y = false);
 			void run(uint8_t* cue);
 	};
 }